+++ CHANGELOG.md
@@ -16,7 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add templates for listing offers
 - Add moderation for the offers
 - add offer edit form
-- template of sharing collection request #831
+- add Request collection for volunteers - #831 #832
+- add Request Archive form - #832
+- add Request Solve form - #830
 
 ### Changed
 - update all Javascript Dependencies